Se agrega parte de control vmBox
